TakeLastTimed with backpressure support
